refactor: 상수 클래스 폴더 분리
