+++ src/main/java/constant/message/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package constant;
+package constant.message;
 
 public enum ExceptionMessage {
