+++ src/main/java/constant/message/OutputViewMessage.java
@@ -1,4 +1,4 @@
-package constant;
+package constant.message;
 
 public enum OutputViewMessage {
